fix: updated some android tests and markers
