+++ tests/android/pages/android_base_page.py
@@ -59,6 +59,9 @@ class AndroidBasePage:
             AppiumBy.ANDROID_UIAUTOMATOR, 'new UiSelector().className("android.widget.ProgressBar")'
         )
         self._text_element_selector = 'new UiSelector().text("{}")'
+        self._accept_cookies_button = Element(
+            AppiumBy.ANDROID_UIAUTOMATOR, 'new UiSelector().resourceId("onetrust-accept-btn-handler")'
+        )
 
     @property
     def sb_search_field(self) -> Element:
@@ -190,6 +193,16 @@ class AndroidBasePage:
         """
 
         return self._ai_assistant_dismiss_button
+
+    @property
+    def accept_cookies_button(self) -> Element:
+        """Get accept cookies button
+
+        Returns:
+            Element: accept cookies button
+        """
+
+        return self._accept_cookies_button
 
     def find_by_text_on_screen(self, text: str, raise_error: bool = True) -> Element | None:
         """Find element by text.
